@@ -33,6 +33,9 @@ android {
         resValue "string", "app_minor_version", "${versionMinor}"
 //        buildConfigField "boolean", "CAN_REQUEST_RATING", "false"
 //        buildConfigField "String", "BUILD_TIME", "\"${buildTime()}\""
+        buildFeatures {
+            buildConfig true
+        }
 
         multiDexEnabled true
         testInstrumentationRunner "org.gnucash.android.test.ui.util.GnucashAndroidTestRunner"
@@ -142,8 +145,8 @@ android {
 
     compileOptions { //we want switch with strings during xml parsing
         encoding "UTF-8"
-        sourceCompatibility JavaVersion.VERSION_1_7
-        targetCompatibility JavaVersion.VERSION_1_7
+        sourceCompatibility JavaVersion.VERSION_1_8
+        targetCompatibility JavaVersion.VERSION_1_8
     }
 
     testOptions {
@@ -212,8 +215,12 @@ dependencies {
                    'com.android.support:multidex:1.0.1'
     )
 
+    implementation "androidx.annotation:annotation:1.8.2"
+
     debugImplementation 'com.facebook.stetho:stetho:1.5.0'
 
+    //TODO Butterknife is deprecated. So it needs to be replaced with
+    // Android's binding APIs.
     implementation 'com.jakewharton:butterknife:8.8.1'
     annotationProcessor 'com.jakewharton:butterknife-compiler:8.8.1'
 
