@@ -18,7 +18,7 @@ package org.gnucash.android.export.csv;
 
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import org.gnucash.android.R;
