@@ -16,7 +16,7 @@
 
 package org.gnucash.android.model;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import org.joda.time.LocalDateTime;
