@@ -30,7 +30,7 @@ import android.inputmethodservice.Keyboard;
 import android.inputmethodservice.KeyboardView;
 import android.inputmethodservice.KeyboardView.OnKeyboardActionListener;
 import android.provider.Settings;
-import android.support.annotation.XmlRes;
+import androidx.annotation.XmlRes;
 import android.text.Editable;
 import android.view.HapticFeedbackConstants;
 import android.view.View;
