@@ -17,7 +17,7 @@ package org.gnucash.android.util;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.support.v4.app.JobIntentService;
 import android.util.Log;
 
