@@ -1,6 +1,6 @@
 package org.gnucash.android.util;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.util.Log;
 
 import org.gnucash.android.export.ExportAsyncTask;
